validator_python/requirements.txt
requests==2.31.0
pytest==7.4.3
python-dotenv==1.0.0
pyyaml==6.0.1
matplotlib==3.8.2
numpy==1.26.3
fastapi==0.109.0
uvicorn==0.27.0 